+++ app/src/main/kotlin/io/github/wykopmobilny/models/dataclass/LinkVoteResponsePublishModel.kt
@@ -3,6 +3,6 @@ package io.github.wykopmobilny.models.dataclass
 import io.github.wykopmobilny.api.responses.DigResponse
 
 data class LinkVoteResponsePublishModel(
-    val linkId: Int,
+    val linkId: Long,
     val voteResponse: DigResponse
 )
